Rename "name_index" to "field_name_index" and "method_name_index", and add bounds checking to attribute parsing
This makes the code for class parsing more readable, as you don't
have to juggle different variables with similar names. This also
fixes a potential problem where the field or method name could shadow
a name in attribute parsing. It also ensures that we parse attributes
of the right size, which prevents us from skipping bytes and getting
off track.