Fix logic issue in getstatic and putstatic
Previously, we tried to compare NameAndTypeInfo structs to find
the correct field when resolving statics. This was logically incorrect,
instead we have to resolve the strings in the current method class first
and then do string comparisons.
These logic issues still exist in getfield and putfield, and at some
point we will need to fix those.