Fix CustomObject class access methods
This fixes a compiler error were we couldn't implement Object::class
for CustomObject. This has been fixed by adding a function to Class,
which requries that it be able to transform itself as an Rc into Rc<dyn Class>,
which allows us to return CustomObject's parameterized class as a
dyn Class.