Make class comparisons safer
In the previous fix, we implemented PartialEq for Class, using the
assumption that if two classes had the same class file then they
must be the same. While this *should* always be the case, it may not
always be true in practice. A much better way is to simply use Rc::ptr_eq()
to make sure that the two class Rcs point to the same class object.
To ensure that this method is used from now on, I commented out
the implementation of PartialEq.