IT COMPILES!
Finally. We have a working system for using native objects and classes.
However, before we get to wrapped up in that we have to remember that
the JVM still doesn't work. We need to completely overhaul Instructions,
and provide implementations for some unimplemented ones.

That being said, this is an important step towards being fast.