Add a new 'speed' folder
This folder (and corresponding tests) will be used to compare our
speed vs openjdk (or whatever is registered as 'java' on a machine).
Currently we are slower than openjdk at computing 1000!, but not by
much. This comes from the fact that we don't optimize usercode yet,
so this will get better. It will also get better when we parse code
into Instructions.