Allocate local variables and stack next to each other
This will hopefully speed the JVM up, because the local variables
and the stack will be close in memory, so the CPU can hopefully
cache both of them. I don't actually know if this is true, so more
testing is needed.
We also now have a problem where we segfault running speed/Primes.
For some reason, this doesn't happen with any other file in our test
suite as far as I can tell.
This commit also adds a bunch of #[cfg]s so that we can build on wasm,
enabling website integration.