Fix a segfault in speed/Primes
This patch fixes a segfault introduced by attempting to allocate
the local variables and operand stack of a frame as a single vector.
When the program attempted to drop a frame with a non-empty operand
stack, it convered Values into VarValues, which caused a segfault.
The solution was to simply empty the operand stack (dropped its elements
as Values) before deallocating actual vectors.